+++ src/test/java/cart/controller/AbstractProductControllerTest.java
@@ -1,6 +1,8 @@
 package cart.controller;
 
 import cart.service.ProductSearchService;
+import cart.service.ProductUpdateService;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
@@ -11,6 +13,10 @@ public class AbstractProductControllerTest {
 
     @MockBean
     protected ProductSearchService productSearchService;
+    @MockBean
+    protected ProductUpdateService productUpdateService;
+    @Autowired
+    protected ObjectMapper objectMapper;
     @Autowired
     protected MockMvc mockMvc;
 }
